fix: Internal UAT Feedback
1. edit guru mengubah matpel masih eror
2. notif ubah mapel ditambah padahal engga
3. edit guru ada yg error apapun itu yg di edit
4. pada tambah siswa pada kelas term cancel ubah menjadi batal
5. menghapus siswa notifnya siswa berhasil berhasil dihapus
6. hapus materi dan tugas error terutama yang udah ada aktifitas dari siswa
7. term submit di beri nilai ubah jadi beri nilai

+++ app/Controllers/SessionItem.php
@@ -104,8 +104,11 @@ class SessionItem extends BaseController
     {
         $sessionItem = $this->sessionItemModel->find($session_item_id);
 
-        $filepath = ROOTPATH . 'public/files/session-items/' . $sessionItem['file'];
-        unlink($filepath);
+
+        if ($sessionItem['file'] != null) {
+            $filepath = ROOTPATH . 'public/files/session-items/' . $sessionItem['file'];
+            unlink($filepath);
+        }
 
         $result = $this->sessionItemModel->delete($session_item_id);
         if ($result) {
